requirements.txt
asgiref==3.4.1
Django==3.2.9
django-extensions==3.1.5
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
pytz==2021.3
setuptools==58.3.0
sqlparse==0.4.2
wheel==0.37.0
Pillow==8.4.0